backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-chroma
langchain-text-splitters
chromadb
pymupdf
tiktoken
pydantic-settings
